pyproject.toml
[project]
name = "study-bug"
version = "0.1.0"
description = ""
authors = [
    {name = "LightNow0",email = "kimlightnow@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
